Refactor file names for PianoBase to follow React naming conventions
