+++ app/ukulele/ukulele.tsx
@@ -1,8 +1,8 @@
 import React, { useRef, useEffect, useState } from "react";
 import * as Tone from "tone";
 import { Button, Slider, Heading, Box, Card } from "@radix-ui/themes";
-import { PianoUkulele } from "../pianos/piano_ukulele";
-import UkuleleFretboard from "../ukulele-fretboard/ukulele-fretboard"
+import { PianoUkulele } from "../pianos/PianoUkulele";
+import UkuleleFretboard from "../ukulele-fretboard/UkuleleFretboard"
 
 export default function Ukulele() {
   const synthRef = useRef<Tone.PluckSynth | null>(null);
